fix: keep toString() from bombing on node 6.1.0
This is to address https://github.com/allenluce/node-autovivify/issues/1

Some change occurred between the node repo's commits 893835539 and
079973b96 (corresponding to v8 versions 4.8.271.17 and 4.9.385.18) that
makes toString() cause node to die with a `basic_string::_S_construct
NULL not valid` error message.

It seems that the new node.js calls the property getter with an invalid
property address to query (after first calling for a toString property
on instance and object). This causes the fault.

It's not 100% clear what's happening, but adding a toString() method
gets some functionality working. Using Object.prototype.toString.call()
on the object still bombs.