fix: make Object.prototype.toString.call() work
So when Node 6+ wants to do an Object.prototype.toString call, it first
sends a query for a property (the GetToStringTag property, in
particular). Autovivify didn't know how to deal with this query, so
would bomb.

This has the property getter just return if it's asked for a
symbol. This lets v8 fall through to its own stringify routines, so an
appropriate value is emitted ("[object AutoVivify]").

This also removes the explicit toString method as the existing fall-back
one seems to work just fine.

This fixes #1